rename Chain trait to VanityChain, move Chain from flags to chain.rs
